+++ crates/parser/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "parser"
+version = "0.0.0"
+license = "MIT OR Apache-2.0"
+authors = ["rust-analyzer developers"]
+edition = "2018"
+
+[lib]
+doctest = false
+
+[dependencies]
+drop_bomb = "0.1.4"
