+++ crates/parser/src/grammar/paths.rs
