--- crates/ra_parser/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-edition = "2018"
-name = "ra_parser"
-version = "0.1.0"
-authors = ["rust-analyzer developers"]
-publish = false
-license = "MIT OR Apache-2.0"
-
-[lib]
-doctest = false
-
-[dependencies]
-drop_bomb = "0.1.4"
